# Improve performance by reducing allocations