--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,11 @@
+module github.com/hokaccha/go-prettyjson
+
+go 1.18
+
+require github.com/fatih/color v1.14.0
+
+require (
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.17 // indirect
+	golang.org/x/sys v0.4.0 // indirect
+)