This PR improves the performance by reducing allocations of strings. This is a recreate of #17. Now that https://github.com/fatih/color/issues/124 is resolved and we can set the color without `unsafe` package, I'd like to suggest performance improvement again after 2.5 years.
```
name      old time/op    new time/op    delta
Fromat-8    6.43µs ±  0%    3.01µs ±  0%  -53.28%  (p=0.000 n=9+9)

name      old alloc/op   new alloc/op   delta
Fromat-8    6.00kB ±  0%    2.92kB ±  0%  -51.36%  (p=0.000 n=10+10)

name      old allocs/op  new allocs/op  delta
Fromat-8       191 ±  0%        50 ±  0%  -73.82%  (p=0.000 n=10+10)
```